backend/app/storage/extracted/project-root/project-root/backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pytest
requests
python-dotenv
numpy